apps/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "usf-api"
version = "0.1.0"
description = "USF API Gateway"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlmodel>=0.0.19",
    "psycopg[async]>=3.2",
    "loguru>=0.7",
    "httpx>=0.27",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "redis>=5.0",
    "pydantic-settings>=2.0",
    "sqlalchemy[asyncio]>=2.0",
    "usf-core",
]

[tool.hatch.build.targets.wheel]
packages = ["usf_api"]
